pyproject.toml
[build-system]
requires = [
    "setuptools>=42",
    "wheel",
]
build-backend = "setuptools.build_meta"

[project]
name = "mace-torch"
version = "0.3.8"
description = "MACE - Fast and accurate machine learning interatomic potentials with higher order equivariant message passing."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "torch>=1.12",
    "e3nn==0.4.4",
    "numpy",
    "opt_einsum",
    "ase",
    "torch-ema",
    "prettytable",
    "matscipy",
    "h5py",
    "torchmetrics",
    "python-hostlist",
    "configargparse",
    "GitPython",
    "pyYAML",
    "tqdm",
    "lmdb",
    "orjson",
    "matplotlib",
    "pandas",
    "nvidia-ml-py",
    "kagglehub",
    "einops",
]

# Make isort compatible with black
[tool.isort]
profile = "black"

# Pylint
[tool.pylint.'MESSAGES CONTROL']
disable = [
    "line-too-long",
    "no-member",
    "missing-module-docstring",
    "missing-class-docstring",
    "missing-function-docstring",
    "too-many-arguments",
    "too-many-positional-arguments",
    "too-many-locals",
    "too-many-return-statements",
    "not-callable",
    "logging-fstring-interpolation",
    "logging-not-lazy",
    "logging-too-many-args",
    "invalid-name",
    "too-few-public-methods",
    "too-many-instance-attributes",
    "too-many-statements",
    "too-many-branches",
    "import-outside-toplevel",
    "cell-var-from-loop",
    "duplicate-code",
    "use-dict-literal",
]

[tool.pylint.MASTER]
ignore-paths = [
    "^mace/tools/torch_geometric/.*$",
    "^mace/tools/scatter.py$",
]

[tool.pylint.FORMAT]
max-module-lines = 2000
